Reworked SENSEI interface to share more code with LibSim
