+++ lib/vistle/insitu/core/CMakeLists.txt
@@ -1,5 +1,12 @@
-set(SOURCES exception.cpp slowMpi.cpp)
-set(HEADER callFunctionWithVoidToTypeCast.h dataType.h exception.h export.h slowMpi.h transformArray.h)
+set(SOURCES exception.cpp slowMpi.cpp attachVistleShm.cpp)
+set(HEADER
+    callFunctionWithVoidToTypeCast.h
+    dataType.h
+    exception.h
+    export.h
+    slowMpi.h
+    transformArray.h
+    attachVistleShm.h)
 
 vistle_add_library(vistle_insitu_core ${SOURCES} ${HEADER})
 vistle_export_library(vistle_insitu_core ${SOURCES} ${HEADER})
@@ -7,13 +14,10 @@ vistle_target_link_libraries(
     vistle_insitu_core
     PRIVATE
     Boost::boost
-    PRIVATE
     Boost::serialization
-    PRIVATE
     MPI::MPI_C
-    PRIVATE
     ${BOOST_MPI}
-    PRIVATE
     vistle_util
+    vistle_util_mpi
     PUBLIC
     vistle_core)
