+++ lib/vistle/insitu/message/moduleInfo.cpp
@@ -7,10 +7,6 @@ bool ModuleInfo::isInitialized() const
 {
     return m_initialized;
 }
-bool ModuleInfo::isReady() const
-{
-    return m_ready;
-}
 bool ModuleInfo::isPortConnected(const std::string &portName) const
 {
     return m_connectedPorts.find(portName) != m_connectedPorts.end();
@@ -59,9 +55,6 @@ bool ModuleInfo::update(const Message &msg)
         auto m = msg.unpackOrCast<message::DisconnectPort>();
         m_connectedPorts.erase(m.value);
     } break;
-    case InSituMessageType::Ready: {
-        m_ready = msg.unpackOrCast<Ready>().value;
-    } break;
     case InSituMessageType::ShmInfo: {
         m_shmInfo = msg.unpackOrCast<message::ShmInfo>().value;
     } break;
@@ -84,8 +77,4 @@ void ModuleInfo::update(const ModuleInfo::ShmInfo &shmInfo)
 void ModuleInfo::setInitState(bool state)
 {
     m_initialized = state;
-}
-void ModuleInfo::setReadyState(bool state)
-{
-    m_ready = state;
 }
