+++ lib/vistle/insitu/sensei/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(SENSEI_SOURCES metaData.cpp objectRetriever.cpp sensei.cpp)
-set(SENSEI_HEADER exception.h export.h intOption.h metaData.h objectRetriever.h sensei.h)
+set(SENSEI_HEADER exception.h export.h metaData.h objectRetriever.h sensei.h)
 
 vistle_add_library(vistle_sensei ${SENSEI_SOURCES} ${SENSEI_HEADER})
 vistle_export_library(vistle_sensei ${SENSEI_SOURCES} ${SENSEI_HEADER})
