+++ lib/vistle/insitu/sensei/sensei.cpp
@@ -1,13 +1,15 @@
 #include "sensei.h"
 #include "exception.h"
-#include "intOption.h"
 
+#include <vistle/insitu/core/attachVistleShm.h>
 #include <vistle/insitu/message/ShmMessage.h>
 #include <vistle/insitu/message/SyncShmIDs.h>
 #include <vistle/insitu/message/addObjectMsq.h>
+#include <vistle/insitu/message/sharedParam.h>
+
 #include <vistle/util/directory.h>
-#include <vistle/util/filesystem.h>
 #include <vistle/util/enumarray.h>
+#include <vistle/util/filesystem.h>
 #include <vistle/util/hostname.h>
 #include <vistle/util/shmconfig.h>
 
@@ -32,15 +34,23 @@ namespace vistle {
 namespace insitu {
 namespace sensei {
 namespace detail {
+
+int getRank(MPI_Comm Comm)
+{
+    int rank = 0;
+    MPI_Comm_rank(Comm, &rank);
+    return rank;
+}
+
 struct Internals {
+    Internals(int rank): messageHandler(rank) {}
     message::ModuleInfo moduleInfo;
     std::unique_ptr<vistle::insitu::message::AddObjectMsq>
         sendMessageQueue; // Queue to send addObject messages to module
 
     message::InSituShmMessage messageHandler;
     // options that can be set in the module
-    EnumArray<IntOption, IntOptions> intOptions{IntOption{IntOptions::NthTimestep, 1},
-                                                IntOption{IntOptions::KeepTimesteps, true}};
+    std::vector<message::IntParam> moduleParams{{"frequency", 1}, {"keep timesteps", true}};
 };
 } // namespace detail
 } // namespace sensei
@@ -49,19 +59,15 @@ struct Internals {
 
 SenseiAdapter::SenseiAdapter(bool paused, MPI_Comm Comm, MetaData &&meta, ObjectRetriever cbs,
                              const std::string &options)
-: m_callbacks(cbs), m_metaData(std::move(meta)), m_internals(new detail::Internals{})
+: m_callbacks(cbs), m_metaData(std::move(meta)), m_internals(new detail::Internals{detail::getRank(Comm)})
 {
     MPI_Comm_rank(Comm, &m_rank);
     MPI_Comm_size(Comm, &m_mpiSize);
     m_commands["run/paused"] = !paused; // if true run else wait
     m_commands["exit"] = false; // let the simulation know that vistle wants to
         // exit by returning false from execute
-    try {
-        m_internals->messageHandler.initialize(m_rank);
-        dumpConnectionFile(comm);
-    } catch (...) {
-        throw Exception() << "failed to create connection facilities for Vistle";
-    }
+    dumpConnectionFile(comm);
+
 #ifdef MODULE_THREAD
     startVistle(comm, options);
 #endif
@@ -141,6 +147,7 @@ bool SenseiAdapter::quitRequested()
 {
     if (m_commands["exit"]) {
         m_internals->messageHandler.send(ConnectionClosed{true});
+        insitu::detachShm();
         return true;
     }
     return false;
@@ -163,7 +170,7 @@ bool SenseiAdapter::WaitedForModuleCommands()
 
 bool SenseiAdapter::haveToProcessTimestep(size_t timestep)
 {
-    return timestep % m_internals->intOptions[IntOptions::NthTimestep].value() == 0;
+    return timestep % message::getIntParamValue(m_internals->moduleParams, "frequency");
 }
 
 void SenseiAdapter::processData()
@@ -179,7 +186,7 @@ void SenseiAdapter::processData()
         updateMeta(dataObject.object());
         addObject(dataObject.portName(), dataObject.object());
     }
-    if (m_internals->intOptions[IntOptions::KeepTimesteps].value())
+    if (message::getIntParamValue(m_internals->moduleParams, "Keep timesteps"))
         ++m_processedTimesteps;
     else
         ++m_iterations;
@@ -250,9 +257,6 @@ bool SenseiAdapter::recvAndHandeMessage(bool blocking)
     case InSituMessageType::DisconnectPort: {
         calculateUsedData();
     } break;
-    case InSituMessageType::Ready: {
-        m_processedTimesteps = 0;
-    } break;
     case InSituMessageType::ExecuteCommand: {
         ExecuteCommand exe = msg.unpackOrCast<ExecuteCommand>();
         auto it = m_commands.find(exe.value.first);
@@ -265,12 +269,13 @@ bool SenseiAdapter::recvAndHandeMessage(bool blocking)
     case InSituMessageType::ConnectionClosed: {
         m_internals->sendMessageQueue.reset(nullptr);
         m_connected = false;
+        insitu::detachShm();
         CERR << "connection closed" << endl;
     } break;
-    case InSituMessageType::SenseiIntOption: {
-        auto option = msg.unpackOrCast<SenseiIntOption>().value;
-        update(m_internals->intOptions, option);
-        if (option.name() == sensei::IntOptions::KeepTimesteps) {
+    case InSituMessageType::IntOption: {
+        auto option = msg.unpackOrCast<IntOption>().value;
+        updateIntParam(m_internals->moduleParams, option);
+        if (option.name == "keep timesteps") {
             m_processedTimesteps = 0;
             ++m_iterations;
         }
@@ -346,16 +351,7 @@ bool SenseiAdapter::initializeVistleEnv()
 
 void SenseiAdapter::initializeMessageQueues() throw()
 {
-    auto shmName = m_internals->moduleInfo.shmName();
-    CERR << "attaching to shm: name = " << shmName << " id = " << m_internals->moduleInfo.id() << " rank = " << m_rank
-         << endl;
-    bool shmPerRank = vistle::shmPerRank();
-    vistle::Shm::attach(shmName, m_internals->moduleInfo.id(), m_rank, shmPerRank);
-    static int numCalls = 0;
-    ++numCalls;
-    std::string suffix = "_sim";
-    suffix += numCalls;
-    vistle::Shm::setExternalSuffix(suffix);
+    insitu::attachShm(m_internals->moduleInfo.shmName(), m_internals->moduleInfo.id(), m_rank);
     m_internals->sendMessageQueue.reset(new AddObjectMsq(m_internals->moduleInfo, m_rank));
 }
 
@@ -408,12 +404,5 @@ void SenseiAdapter::updateMeta(vistle::Object::ptr obj) const
         obj->setExecutionCounter(m_executionCount);
         obj->setTimestep(m_processedTimesteps);
         obj->setIteration(m_iterations);
-        CERR << "adding object for timestep " << m_processedTimesteps << " on iteration " << m_iterations << std::endl;
     }
-}
-
-
-bool SenseiAdapter::isMOduleReady()
-{
-    return m_internals->moduleInfo.isReady();
 }
